+++ src/models/operations/getchannelschannelhashedidchannelepisodeschannelepisodeid.ts
@@ -37,18 +37,12 @@ export type GetChannelsChannelHashedIdChannelEpisodesChannelEpisodeIdCode =
     typeof GetChannelsChannelHashedIdChannelEpisodesChannelEpisodeIdCode
   >;
 
-/**
- * The type of episode.
- */
 export const GetChannelsChannelHashedIdChannelEpisodesChannelEpisodeIdEpisodeType =
   {
     Full: "full",
     Trailer: "trailer",
     Bonus: "bonus",
   } as const;
-/**
- * The type of episode.
- */
 export type GetChannelsChannelHashedIdChannelEpisodesChannelEpisodeIdEpisodeType =
   ClosedEnum<
     typeof GetChannelsChannelHashedIdChannelEpisodesChannelEpisodeIdEpisodeType
@@ -67,15 +61,16 @@ export type GetChannelsChannelHashedIdChannelEpisodesChannelEpisodeIdPodcastSett
      */
     episodeType?:
       | GetChannelsChannelHashedIdChannelEpisodesChannelEpisodeIdEpisodeType
+      | null
       | undefined;
     /**
      * The number of the episode.
      */
-    episodeNumber?: number | undefined;
+    episodeNumber?: number | null | undefined;
     /**
      * The season number of the episode.
      */
-    seasonNumber?: number | undefined;
+    seasonNumber?: number | null | undefined;
     /**
      * Whether the episode contains explicit content.
      */
@@ -208,11 +203,11 @@ export const GetChannelsChannelHashedIdChannelEpisodesChannelEpisodeIdPodcastSet
     z.ZodTypeDef,
     unknown
   > = z.object({
-    episode_type:
-      GetChannelsChannelHashedIdChannelEpisodesChannelEpisodeIdEpisodeType$inboundSchema
-        .optional(),
-    episode_number: z.number().int().optional(),
-    season_number: z.number().int().optional(),
+    episode_type: z.nullable(
+      GetChannelsChannelHashedIdChannelEpisodesChannelEpisodeIdEpisodeType$inboundSchema,
+    ).optional(),
+    episode_number: z.nullable(z.number().int()).optional(),
+    season_number: z.nullable(z.number().int()).optional(),
     explicit_content: z.boolean().optional(),
     hide_from_feed: z.boolean().optional(),
   }).transform((v) => {